LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/view-timeline-with-transform-on-subject-expected.txt


FAIL ViewTimeline use untransformed box for range calculations assert_approx_equals: progress at contain 100% expected 1 +/- 0.000001 but got 0
